+++ g2p_registrant/models/reg_relationship.py
@@ -22,9 +22,10 @@ class G2PRegistrantRelationship(models.Model):
     _name = 'g2p.reg.rel'
     _description = 'Registrant Relationship'
     _order = 'id desc'
+    _inherit = ['mail.thread']
 
-    registrant1 = fields.Many2one('res.partner','Registrant 1')
-    registrant2 = fields.Many2one('res.partner','Registrant 2')
+    registrant1 = fields.Many2one('res.partner','Registrant 1', required=True, domain=[('is_registrant','=',True),('is_group','=',False)])
+    registrant2 = fields.Many2one('res.partner','Registrant 2', required=True, domain=[('is_registrant','=',True),('is_group','=',False)])
     relation = fields.Many2one('g2p.relationship','Relation')
     disabled = fields.Datetime('Date Disabled')
     disabled_by = fields.Many2one('res.users', 'Disabled by')
@@ -48,6 +49,22 @@ class G2PRegistrantRelationship(models.Model):
         if name:
             args = ['|',('registrant1', operator, name),('registrant2', operator, name)] + args
         return self._search(args, limit=limit, access_rights_uid=name_get_uid)
+
+    def disable_relationship(self):
+        for rec in self:
+            if not rec.disabled:
+                rec.update({
+                    'disabled':fields.Datetime.now(),
+                    'disabled_by':self.env.user,
+                })
+
+    def enable_relationship(self):
+        for rec in self:
+            if rec.disabled:
+                rec.update({
+                    'disabled':None,
+                    'disabled_by':None,
+                })
 
 class G2PRelationship(models.Model):
     _name = 'g2p.relationship'