+++ g2p_registrant/models/registrant.py
@@ -22,7 +22,6 @@ class G2PRegistrant(models.Model):
     _inherit = 'res.partner'
 
     #Custom Fields
-    #Todo: location = fields.Many2one('g2p.location'), in location module
     address = fields.Text('Address', tracking=True)
     #addl_fields = fields.Text('Additional Fields', tracking=True)
     disabled = fields.Datetime('Date Disabled', tracking=True)
@@ -30,6 +29,7 @@ class G2PRegistrant(models.Model):
     disabled_by = fields.Many2one('res.users', 'Disabled by', tracking=True)
     #Tag (exist in res.partner: category_id)
     
+    reg_ids = fields.One2many('g2p.reg.id','registrant','Registrant IDs')
     is_registrant = fields.Boolean('Registrant')
     is_group = fields.Boolean('Group')